First version of crcany.c, which computes any CRC several ways.
crcany.c computes any CRC up to 128 bits long (if a 64-bit integer
type is available), and does so using bit-wise, byte-wise, and
word-wise algorithms, adjusting as needed for the endianess of the
architecture it is running on. The byte-wise and word-wise CRC
calculations are only done for CRCs up to 64 bits in length.