Move common code to model.c and permit negative numbers.
A large amount of duplicated code in crcany.c and mincrc.c was
moved to model.h and model.c. This facilitated enhancing that
common code to permit negative numbers on the input. Now values
like 0xffffffff can be represented as -1 instead. mincrc.c was
updated to make use of negative when it makes the output smaller.