Avoid use of uintmax_t outside loop for little endian.
This makes the word-wise tables smaller for little-endian, and
permits the by-word loop to use the smaller CRC integer type.
This was considered for big-endian as well, but the additional
shift operation in each by-word loop slowed down the calculation,
even though the smaller tables would take less cache space.