Add creation of remaining bits function to crcgen.
crc*_rem() functions are generated to compute the associated CRC
on 0..8 bits. This permits the computation of a CRC on a message
with a number of bits that is not a mutiple of eight. This
function would normally be applied to the last 1..7 bits of the
message. However it will work properly for 0 or 8 bits as well.