Add options to crcadd for endianess and word size.
crcadd defaults to the endianess and largest integer size of the
machine and compiler it is running on. With the options, the code
can be cross-generated for another machine with a different
endianess, and with a smaller word size (32 bits instead of 64
bits), if desired. These options only affect the generated code
and tables for the word-wise calculation.