Prefer -1 in mincrc for parameters.
Use -1 always when the parameter can be represented as -1 for the
given CRC width. Since 0 and all 1's are the most common values
for init and xorout, this canoncalizes them to be 0 and -1 in
those cases. This actually violates the "minimum" objective of
mincrc for width=3, and increases the length of allcrcs-abbrev.txt
by two bytes, one for each of two CRC-3's. But I like it better
this way.